Remove unnecessary (and glitchy) async=false
This seemed to interact poorly with the other streams when testing.
Initially, this worked great on get, but it totally broke the stream on
stay, which I isolated to the presence of an additional rtmp sink in the
output. I think what was happening was the different muxers ahead of
each sink were out of step and deadlocking the pipeline -- adding an
unlimited size multiqueue seemed to help.

Having streamdelay start with 0 queueing threshold is a better fix,
because removing the async=false stopped the glitchy stalling behavior.