Give each sink an individual parser
The RTMP audio was out of sync and YouTube audio had artifacts. If I
rewired the SRT sink to play the RTMP audio/video, the pipeline would
stall. Removing queues would cause the pipeline to fail to initialize
with "Internal data stream error". I'm learning that it's necessary to
have a distinct parser element in front of each muxer. Not 100% sure
what is happening internally, but it seems like the parsers are
affecting the timestamping, and it matters that sink flows that are not
in lock step parse independently.